Permit setting a custom ID generator
We have a use case where we're having the Trace ID serve dual purpose:
First, for tracing; secondly, to maintain an audit log of data change events.
To support the second use case, we'd like to generate our own ID's
in accordance to the KSUID standard published by Segment.io, so that our
ID's are time-sortable. Rather than pull the library in here, I instead
propose a configuration option on the Tracer that would allow us to override
the Generator with one of our own.